remove set prefixes in params.go and move to e2e_setup_test.go
